requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]==3.12
scikit-learn
mlflow
pyaml
dagshub
fastapi
uvicorn
python-multipart
seaborn
prometheus_client

#-e .